BUG: El kernel del filtre gaussià no es calculava del tot correctament. A més, s'ha convertit el radi en un enter.

+++ starviewer/src/extensions/playground/experimental3d/cudafiltering.h
@@ -13,7 +13,7 @@ extern "C" {
 #endif
 
 
-QVector<float> cfGaussianDifference( vtkImageData *image, float radius );
+QVector<float> cfGaussianDifference( vtkImageData *image, int radius );
 QVector<float> cfBoxMeanDifference( vtkImageData *image, int radius );
